move native tests to runner
