+++ jitexecutor/jitexecutor-dmn/src/main/java/org/kie/kogito/jitexecutor/dmn/SchemaResource.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.jitexecutor.dmnexecutor;
+package org.kie.kogito.jitexecutor.dmn;
 
 import java.io.StringReader;
 import java.util.Arrays;
