+++ jitexecutor/jitexecutor-dmn/src/main/resources/META-INF/native-image/org.kie.kogito/jitexecutor-dmn/reflect-config.json
