+++ jitexecutor/jitexecutor-runner/pom.xml
@@ -10,7 +10,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <artifactId>jitexecutor-runner</artifactId>
-  <name>Kogito :: JIT DMN Runner</name>
+  <name>Kogito :: JIT Executor Runner</name>
 
   <dependencies>
     <dependency>
